+++ pkg/registration/hub/gc/controller_test.go
@@ -90,7 +90,7 @@ func TestGController(t *testing.T) {
 				clusterClient,
 				kubeClient,
 				metadataClient,
-				register.NewNoopApprover(),
+				register.NewNoopHubDriver(),
 				events.NewInMemoryRecorder(""),
 				[]string{"addon.open-cluster-management.io/v1alpha1/managedclusteraddons",
 					"work.open-cluster-management.io/v1/manifestworks"},
@@ -116,7 +116,7 @@ func TestGController(t *testing.T) {
 						kubeInformerFactory.Rbac().V1().ClusterRoles().Lister(),
 						kubeInformerFactory.Rbac().V1().RoleBindings().Lister(),
 						workInformerFactory.Work().V1().ManifestWorks().Lister(),
-						register.NewNoopApprover(),
+						register.NewNoopHubDriver(),
 						events.NewInMemoryRecorder(""),
 						true),
 				},
