+++ pkg/registration/hub/managedcluster/controller.go
@@ -44,7 +44,6 @@ type managedClusterController struct {
 	clusterLister listerv1.ManagedClusterLister
 	applier       *apply.PermissionApplier
 	patcher       patcher.Patcher[*v1.ManagedCluster, v1.ManagedClusterSpec, v1.ManagedClusterStatus]
-	approver      register.Approver
 	hubDriver     register.HubDriver
 	eventRecorder events.Recorder
 }
@@ -58,7 +57,6 @@ func NewManagedClusterController(
 	clusterRoleInformer rbacv1informers.ClusterRoleInformer,
 	rolebindingInformer rbacv1informers.RoleBindingInformer,
 	clusterRoleBindingInformer rbacv1informers.ClusterRoleBindingInformer,
-	approver register.Approver,
 	hubDriver register.HubDriver,
 	recorder events.Recorder) factory.Controller {
 
@@ -66,7 +64,6 @@ func NewManagedClusterController(
 		kubeClient:    kubeClient,
 		clusterClient: clusterClient,
 		clusterLister: clusterInformer.Lister(),
-		approver:      approver,
 		hubDriver:     hubDriver,
 		applier: apply.NewPermissionApplier(
 			kubeClient,
@@ -117,7 +114,13 @@ func (c *managedClusterController) sync(ctx context.Context, syncCtx factory.Syn
 		// If the ManagedClusterAutoApproval feature is enabled, we automatically accept a cluster only
 		// when it joins for the first time, afterwards users can deny it again.
 		if _, ok := managedCluster.Annotations[clusterAcceptedAnnotationKey]; !ok {
-			return c.acceptCluster(ctx, managedCluster)
+			acceptCluster, err := c.hubDriver.Accept(ctx, managedCluster)
+			if err != nil {
+				return err
+			}
+			if acceptCluster {
+				return c.acceptCluster(ctx, managedCluster)
+			}
 		}
 	}
 
@@ -159,7 +162,7 @@ func (c *managedClusterController) sync(ctx context.Context, syncCtx factory.Syn
 			return aggErr
 		}
 
-		if err = c.approver.Cleanup(ctx, managedCluster); err != nil {
+		if err = c.hubDriver.Cleanup(ctx, managedCluster); err != nil {
 			return err
 		}
