+++ pkg/registration/register/csr/hub_driver_test.go
@@ -336,14 +336,67 @@ func TestNewApprover(t *testing.T) {
 	informerFactory := informers.NewSharedInformerFactory(kubeClient, 3*time.Minute)
 	recorder := eventstesting.NewTestingEventRecorder(t)
 	utilruntime.Must(features.HubMutableFeatureGate.Add(ocmfeature.DefaultHubRegistrationFeatureGates))
-	_, err := NewCSRApprover(kubeClient, informerFactory, []string{}, recorder)
+	_, err := NewCSRHubDriver([]string{}, kubeClient, informerFactory, []string{}, recorder)
 	if err != nil {
 		t.Error(err)
 	}
 
 	features.HubMutableFeatureGate.Set(fmt.Sprintf("%s=true", ocmfeature.ManagedClusterAutoApproval))
-	_, err = NewCSRApprover(kubeClient, informerFactory, []string{}, recorder)
+	_, err = NewCSRHubDriver([]string{}, kubeClient, informerFactory, []string{}, recorder)
 	if err != nil {
 		t.Error(err)
 	}
+}
+
+func TestAccept(t *testing.T) {
+
+	cases := []struct {
+		name       string
+		cluster    *clusterv1.ManagedCluster
+		isAccepted bool
+	}{
+		{
+			name: "Accept cluster when annotations not present",
+			cluster: &clusterv1.ManagedCluster{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "managed-cluster1",
+				},
+			},
+			isAccepted: true,
+		},
+		{
+			name: "Reject cluster when aws irsa annotations are present.",
+			cluster: &clusterv1.ManagedCluster{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "managed-cluster2",
+					Annotations: map[string]string{
+						"agent.open-cluster-management.io/managed-cluster-arn":             "arn:aws:eks:us-west-2:123456789012:cluster/managed-cluster2",
+						"agent.open-cluster-management.io/managed-cluster-iam-role-suffix": "7f8141296c75f2871e3d030f85c35692",
+					},
+				},
+			},
+			isAccepted: false,
+		},
+	}
+
+	kubeClient := kubefake.NewSimpleClientset()
+	informerFactory := informers.NewSharedInformerFactory(kubeClient, 3*time.Minute)
+	recorder := eventstesting.NewTestingEventRecorder(t)
+	utilruntime.Must(features.HubMutableFeatureGate.Add(ocmfeature.DefaultHubRegistrationFeatureGates))
+	csrHubDriver, err := NewCSRHubDriver([]string{}, kubeClient, informerFactory, []string{}, recorder)
+
+	if err != nil {
+		t.Error(err)
+	}
+
+	for _, c := range cases {
+		t.Run(c.name, func(t *testing.T) {
+			isAccepted, err := csrHubDriver.Accept(context.TODO(), c.cluster)
+			if err == nil && c.isAccepted != isAccepted {
+				t.Errorf("expect %t, but %t", c.isAccepted, isAccepted)
+			}
+		},
+		)
+	}
+
 }
