+++ pkg/registration/register/interface.go
@@ -76,23 +76,19 @@ type RegisterDriver interface {
 	ManagedClusterDecorator(cluster *clusterv1.ManagedCluster) *clusterv1.ManagedCluster
 }
 
-// Approvers is the inteface that each driver should implement on hub side. The hub controller will use this driver
-// to check the registration request from the agent and cleanup.
-type Approver interface {
-	// Run starts a reconciler on the hub side to monitor the registration request and approve the request
-	// if necessary. This is a blocking call.
-	Run(ctx context.Context, workers int)
-
-	// Cleanup is executed when hubAcceptClient in ManagedCluster is set false or cluster is deleting. The hub controller
-	// deletes rolebindings for the agent, and then this is the additional operation a driver should process.
-	Cleanup(ctx context.Context, cluster *clusterv1.ManagedCluster) error
-}
-
 // HubDriver interface is used to implement operations required to complete aws-irsa registration and csr registration.
-// The Approver interface above is used to implement operations related to approving the CSR request, and permission
-// creation is not related to CSR approval. Hence, created CreatePermissions under a separate interface.
+// The hub controller will use this driver to check the registration request from the agent and cleanup.
 type HubDriver interface {
 	// CreatePermissions is executed when hubAcceptClient in ManagedCluster is set to true. The hub controller creates the
 	// required permissions for the spoke to be able to access resources on the hub cluster.
 	CreatePermissions(ctx context.Context, cluster *clusterv1.ManagedCluster) error
+	// Run starts a reconciler on the hub side to monitor the registration request and approve the request
+	// if necessary. This is a blocking call.
+	Run(ctx context.Context, workers int)
+	// Cleanup is executed when hubAcceptClient in ManagedCluster is set false or cluster is deleting. The hub controller
+	// deletes rolebindings for the agent, and then this is the additional operation a driver should process.
+	Cleanup(ctx context.Context, cluster *clusterv1.ManagedCluster) error
+	// Accept is executed when autoapproval is enabled to verify if the managedcluster uses csr or awsirsa and
+	// setting hubAcceptClient based on autoApprovedIdentities
+	Accept(ctx context.Context, cluster *clusterv1.ManagedCluster) (bool, error)
 }
