+++ test/integration/registration/integration_suite_test.go
@@ -209,7 +209,10 @@ var _ = ginkgo.BeforeSuite(func() {
 		go func() {
 			m := hub.NewHubManagerOptions()
 			m.ImportOption.APIServerURL = cfg.Host
+			m.EnabledRegistrationDrivers = []string{"csr", "awsirsa"}
+			m.HubClusterArn = "arn:aws:eks:us-west-2:123456789012:cluster/hub-cluster1"
 			m.ClusterAutoApprovalUsers = []string{util.AutoApprovalBootstrapUser}
+			m.AutoApprovalAwsPatterns = []string{"managed-cluster1"}
 			err := m.RunControllerManager(ctx, &controllercmd.ControllerContext{
 				KubeConfig:    cfg,
 				EventRecorder: util.NewIntegrationTestEventRecorder("hub"),
